requirements.txt
playwright==1.58.0
behave==1.3.3
behave-html-pretty-formatter==1.16.1
python-dotenv==1.2.1